✓ update extendmodule interface (cont.) / Adds SPFM Light
git-svn-id: http://amethyst.yui.ne.jp/svn/pc98/np2/branches/yui@1884 7c2116fe-53d0-4aec-ab59-80607f05abc0

+++ WORK_01/sdl2/ext/spfm/spfmlight.cpp
@@ -44,7 +44,7 @@ bool CSpfmLight::Initialize()
 		OEMCHAR* lpKeyName = szSections;
 		while ((!bOpened) && (*lpKeyName != '\0'))
 		{
-			const UINT cchKeyName = OEMSTRLEN(lpKeyName);
+			const size_t cchKeyName = OEMSTRLEN(lpKeyName);
 			if (milstr_memcmp(lpKeyName, OEMTEXT("SPFM Light")) == 0)
 			{
 				if ((lpKeyName[10] == '(') && (lpKeyName[cchKeyName - 1] == ')'))
